test/pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets==3.0.1",
    "evadb==0.3.9",
    "matplotlib==3.9.2",
    "numpy==2.0.2",
    "opencv-python==4.10.0.84",
    "pandas==2.2.3",
    "pillow==11.0.0",
    "psycopg2-binary==2.9.9",
    "ptflops>=0.7.4",
    "scikit-learn==1.5.2",
    "sentencepiece==0.2.0",
    "thop>=0.1.1.post2209072238",
    "torch==2.4.1",
    "torchvision==0.19.1",
    "transformers==4.45.2",
]
